refactor(go-template): extract type codegen into type/type-codegen.ts
Move TypeScript-type → Go-type rendering out of the adapter class into
pure free functions (Roadmap B): typeInfoToGo, tsTypeStringToGo,
inferTypeFromValue. They read only state.localTypeNames.

typeInfoToGo is no longer a GoEmitContext entry point — it's a free
function now, so value-lowering imports it directly instead of calling
back through the seam, shrinking the context surface to the genuine
recursive entry points plus extractPropNameFromInitialValue.

Output byte-identical, verified by the adapter unit (552) and
adapter-tests conformance (786) suites; tsgo + biome clean.
